+++ tests/__init__.py
@@ -1,0 +1,7 @@
+import os
+
+_THIS_MODULE = os.path.abspath(__file__)
+TESTS_DIR = os.path.dirname(_THIS_MODULE)
+
+ROOT_DIR = os.path.dirname(TESTS_DIR)
+DOCS_DIR = os.path.join(ROOT_DIR, "docs")