+++ docs/changelog.rst
@@ -11,6 +11,8 @@ Development
 - Turning off dereferencing for the results of distinct query. #2663
 - Add tests against Mongo 5.0 in pipeline
 - Drop support for Python 3.6 (EOL)
+- Bug fix support for PyMongo>=4 to fix "InvalidOperation: Cannot use MongoClient after close"
+  errors.
 
 Changes in 0.24.2
 =================